requirements.txt
aiogram==3.0.0
SQLAlchemy==2.0.20
pydantic==2.3.0
asyncpg==0.28.0
redis==5.0.0
python-dotenv==1.0.0
cachetools==5.3.1
pydantic-settings==2.0.3
fluent.runtime==0.4.0